refactor: refactor files naming conventions
