+++ lib/core/design_system/widgets/auth_background_wrapper.dart
