+++ lib/core/navigate/route_manager.dart
@@ -2,21 +2,21 @@ import 'package:flutter/material.dart';
 import 'package:go_router/go_router.dart';
 import 'package:sellio_mobile/core/navigate/route_args.dart';
 import '../../../presentation/screens/auth/create_account/create_account_screen.dart';
-import '../../../presentation/screens/auth/forgetPassword/confirm_password_screen.dart';
-import '../../../presentation/screens/auth/forgetPassword/forget_password_otp_screen.dart';
-import '../../../presentation/screens/auth/forgetPassword/forget_password_screen.dart';
 import '../../../presentation/screens/auth/login/login_screen.dart';
 import '../../../presentation/screens/auth/signupOTP.dart';
 import '../../../presentation/screens/cart/cart_screen.dart';
-import '../../../presentation/screens/customize_product/CustomizeYourProductScreen.dart';
+import '../../../presentation/screens/customize_product/customize_your_product_screen.dart';
 import '../../../presentation/screens/home/home_screen.dart';
 import '../../../presentation/screens/main/dashboard.dart';
 import '../../../presentation/screens/notification/notification_screen.dart';
-import '../../../presentation/screens/product_details/ProductDetailsScreen.dart';
+import '../../../presentation/screens/product_details/product_details_screen.dart';
 import '../../../presentation/screens/store_details/about_store/about_store.dart';
 import '../../../presentation/screens/store_details/store_details_screen.dart';
 import '../../../presentation/screens/thrift_screen.dart';
 import '../../../ui/screens/account/account_screen.dart';
+import '../../presentation/screens/auth/forget_password/confirm_password_screen.dart';
+import '../../presentation/screens/auth/forget_password/forget_password_otp_screen.dart';
+import '../../presentation/screens/auth/forget_password/forget_password_screen.dart';
 import 'app_routes.dart';
 
 class RouteGenerator {
