+++ lib/domain/entities/store_rating.dart
