+++ lib/presentation/screens/auth/signupOTP.dart
@@ -5,7 +5,7 @@ import 'package:sellio_mobile/core/design_system/themes/sellio_theme_provider.da
 import 'package:sellio_mobile/core/design_system/widgets/cards/otp_card.dart';
 import 'package:sellio_mobile/core/localization/localization_service.dart';
 import '../../../core/design_system/themes/sellio_typography.dart';
-import '../../../core/design_system/widgets/AuthBackgroundWrapper.dart';
+import '../../../core/design_system/widgets/auth_background_wrapper.dart';
 
 
 class ConfirmAccountScreen extends StatefulWidget {
