+++ lib/presentation/screens/customize_product/customize_your_product_screen.dart
@@ -2,7 +2,7 @@ import 'package:flutter/material.dart';
 import 'package:flutter_svg/svg.dart';
 import 'package:sellio_mobile/core/localization/localization_service.dart';
 import '../../../core/design_system/widgets/chip_category.dart';
-import 'widgets/GridProduct.dart';
+import 'widgets/grid_product.dart';
 
 class CustomizeYourProductScreen extends StatefulWidget {
   const CustomizeYourProductScreen({super.key});
