+++ lib/presentation/screens/product_details/product_details_screen.dart
@@ -3,7 +3,7 @@ import 'package:flutter_svg/svg.dart';
 import 'package:sellio_mobile/core/design_system/constants/assets.dart';
 import 'package:sellio_mobile/core/design_system/themes/sellio_theme_provider.dart';
 import 'package:sellio_mobile/core/design_system/widgets/buttons/button.dart';
-import 'package:sellio_mobile/core/design_system/widgets/textField.dart';
+import 'package:sellio_mobile/core/design_system/widgets/text_field.dart';
 import 'package:sellio_mobile/core/localization/localization_service.dart';
 
 import '../../../core/design_system/widgets/sellio_app_bar.dart';
