+++ lib/presentation/screens/store_details/about_store/models/contact_info.dart
