+++ lib/presentation/screens/store_details/about_store/widgets/contact_info_item.dart
@@ -1,7 +1,7 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_svg/svg.dart';
 import 'package:sellio_mobile/core/design_system/themes/sellio_theme_provider.dart';
-import '../models/ContactInfo.dart';
+import '../models/contact_info.dart';
 
 class ContactInfoItem extends StatelessWidget {
   final ContactInfo contactInfo;
