+++ lib/presentation/screens/store_details/cubit/store_details_state.dart
@@ -1,7 +1,7 @@
 import 'package:equatable/equatable.dart';
 import '../../../../../../../domain/entities/store.dart';
 import '../../../../../../../domain/entities/product.dart';
-import '../../../../domain/entities/StoreRating.dart';
+import '../../../../domain/entities/store_rating.dart';
 
 
 sealed class StoreDetailsState extends Equatable {
